+++ graphtransliterator/transliterators/sample/__init__.py
@@ -1,3 +1,0 @@
-class Sample:
-    def init():
-        return __file__